+++ vasttools/utils.py
@@ -26,6 +26,7 @@ from astropy.io import fits
 from astropy.wcs import WCS
 from multiprocessing_logging import install_mp_handler
 from typing import Optional, Union, Tuple, List
+from pathlib import Path
 from mocpy import MOC
 
 # crosshair imports
@@ -664,7 +665,7 @@ def create_moc_from_fits(fits_file: str, max_depth: int = 9) -> MOC:
     if not os.path.isfile(fits_file):
         raise Exception("{} does not exist".format(fits_file))
 
-    with fits.open(fits_file) as vast_fits:
+    with open_fits(fits_file) as vast_fits:
         data = vast_fits[0].data
         if data.ndim == 4:
             data = data[0, 0, :, :]
@@ -703,3 +704,32 @@ def strip_fieldnames(fieldnames: pd.Series) -> pd.Series:
     """
 
     return fieldnames.str.rstrip('A')
+
+def open_fits(fits_path: Union[str, Path], memmap: Optional[bool]=True):
+    """
+    This function opens both compressed and uncompressed fits files.
+    
+    Args:
+        fits_path: Path to the fits file
+        memmap: Open the fits file with mmap.
+    
+    Returns:
+        HDUList loaded from the fits file
+    
+    Raises:
+        ValueError: File extension must be .fits or .fits.fz
+    """
+
+    if type(fits_path) == Path:
+        fits_path = str(fits_path)
+
+    hdul = fits.open(fits_path, memmap=memmap)
+
+    if fits_path.endswith('.fits'):
+        return hdul
+    elif fits_path.endswith('.fits.fz'):
+        return fits.HDUList(hdul[1:])
+    else:
+        raise ValueError("Unrecognised extension for {fits_path}."
+                         "File extension must be .fits or .fits.fz"
+                         )