+++ CHANGELOG.md
@@ -7,7 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## [Unreleased](https://github.com/askap-vast/vast-tools/compare/v2.0.0...HEAD)
 
 #### Added
-
+- Added open_fits function to correctly handle compressed image HDUs [#480](https://github.com/askap-vast/vast-tools/pull/480)
 - Added epoch 41 [#470](https://github.com/askap-vast/vast-tools/pull/470)
 - Added warning when querying corrected data [#465](https://github.com/askap-vast/vast-tools/pull/465)
 - Added epochs 38, 39 and 40 [#466](https://github.com/askap-vast/vast-tools/pull/466)
@@ -40,6 +40,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 #### Changed
 
+- Changed opening of all fits files to use open_fits function [#480](https://github.com/askap-vast/vast-tools/pull/480)
 - Changed default behaviour of Image.measure_coord_pixel_values to have `img=False` [#476](https://github.com/askap-vast/vast-tools/pull/476)
 - Changed `Fields.direction` to be loaded from pickle rather than generated at each call [#431](https://github.com/askap-vast/vast-tools/pull/431)
 - Updated RACS dec limit to +50 [#438](https://github.com/askap-vast/vast-tools/pull/438)
@@ -104,6 +105,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 #### List of PRs
 
+- [#480](https://github.com/askap-vast/vast-tools/pull/480): feat: Added open_fits function to handle compressed image HDUs
 - [#476](https://github.com/askap-vast/vast-tools/pull/476): fix: Fixed missing RMS measurements for non-detections 
 - [#472](https://github.com/askap-vast/vast-tools/pull/472): fix: Remove duplicates in field matching
 - [#470](https://github.com/askap-vast/vast-tools/pull/470): feat: Added epoch 41